+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 
 ## #.#.#
 
+### Changed
+- Tidier collection parameters for specific beamline/experiment writers.
+
+
 ### Fixed
 - Numpy2.0 compatibility - stop using deprecated np.string_ alias for fixed-width bytestrings.
 - External links to eiger meta file in NXdetector and detectorSpecific groups re-enabled.